pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "bridge-ai"
version = "0.1.0"
description = "Monolithic self-play bridge AI research scaffold"
readme = "PLAN.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
  "numpy>=1.26",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "torch>=2.2"
]

[project.optional-dependencies]
dev = ["black", "ruff", "mypy", "types-PyYAML"]
ui = ["streamlit>=1.36"]
modal = ["modal>=0.66"]

[project.scripts]
bridge-ai-train = "bridge_ai.training.train_loop:main"
bridge-ai-selfplay = "bridge_ai.selfplay.runner:main"
bridge-ai-ui = "bridge_ai.ui.streamlit_app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

